Convert conversion function to use doubles
pwmFrequency will be passed as double, so converted parameter to double to avoid loss of precision.

+++ lib/Command_Interpreter.cpp
@@ -123,7 +123,7 @@ void Command_Interpreter_RPi5::initializePins() {
     }
 }
 
-Command_Interpreter_RPi5::ThrusterSpec Command_Interpreter_RPi5::convertPwmValue(int pwmFrequency) {
+Command_Interpreter_RPi5::ThrusterSpec Command_Interpreter_RPi5::convertPwmValue(double pwmFrequency) {
 
     // 1100 - 1464 = negative
     // 1536 - 1900 = positive
